Change "output" to "result"
